consume API to add to book
