+++ src/redux/books/books.js
@@ -1,15 +1,11 @@
 const ADD_BOOK = 'ADD_BOOK';
 const REMOVE_BOOK = 'REMOVE_BOOK';
 const GET_BOOKS = 'GET_BOOKS';
+const BASE_URL = 'https://us-central1-bookstore-api-e63c8.cloudfunctions.net/bookstoreApi';
+const id = 'pvUXP2hLb63eHdU4i44l';
+export const END_POINT = `apps/${id}/books`;
 
-const reducer = (
-  state = [
-    { title: 'Lord of the rings', id: '0', author: 'Prince' },
-    { title: 'Harry potter', id: '1', author: 'Monroe' },
-    { title: 'Cindarella', id: '2', author: 'Halle' },
-  ],
-  action,
-) => {
+const reducer = (state = [], action) => {
   switch (action.type) {
     case GET_BOOKS:
       return Object.keys(action.books).map((book) => ({
@@ -17,11 +13,21 @@ const reducer = (
         ...action.books[book][0],
       }));
 
-    case ADD_BOOK:
+    case ADD_BOOK: {
+      const {
+        id, title, author, category,
+      } = action;
       return [
         ...state,
-        { id: action.id, title: action.title, author: action.author },
+        {
+          id,
+          title,
+          author,
+          category,
+        },
       ];
+    }
+
     case REMOVE_BOOK:
       return [...state.filter((book) => book.id !== action.id)];
     default:
@@ -29,14 +35,41 @@ const reducer = (
   }
 };
 
-export const addBook = (id, title, author) => ({
+const addBook = (id, title, author) => ({
   type: ADD_BOOK,
   id,
   title,
   author,
 });
+
+export const addBookAPI = (itemId, title, author, category) => (dispatch) => {
+  fetch(`${BASE_URL}/${END_POINT}`, {
+    method: 'POST',
+    body: JSON.stringify({
+      item_id: itemId,
+      title,
+      author,
+      category,
+    }),
+    headers: {
+      'Content-type': 'application/json; charset=UTF-8',
+    },
+  }).then((res) => {
+    if (res.ok) {
+      dispatch(addBook(itemId, title, author, category));
+    }
+  });
+};
+
 export const removeBook = (id) => ({ type: REMOVE_BOOK, id });
 
 export const getBooks = (books) => ({ type: GET_BOOKS, books });
+export const asyncGetBook = () => (dispatch) => fetch(`${BASE_URL}/${END_POINT}`, {
+  headers: {
+    'Content-type': 'application/json; charset=UTF-8',
+  },
+})
+  .then((res) => res.json())
+  .then((data) => dispatch(getBooks(data)));
 
 export default reducer;
